requirements.txt
flask==3.0.3
gunicorn==22.0.0
playwright==1.44.0
playwright-stealth==1.0.6
google-cloud-storage==2.17.0
google-cloud-scheduler==2.13.3
selenium==4.29.0
webdriver-manager==4.0.2
pyperclip==1.9.0
Pillow==11.1.0
cryptography==44.0.2
